cleaned 002 test
